// React/app/GerAdv_TS/TipoOrigemSucumbencia/Interfaces/interface.TipoOrigemSucumbencia.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface ITipoOrigemSucumbencia {
// 202501251
    id: number;
 
	nome: string,
}

export interface ITipoOrigemSucumbenciaFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface ITipoOrigemSucumbenciaIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}